+++ CHANGES.md
@@ -3,6 +3,7 @@
 * Improve automated parameter extraction (#9)
 * Handle notebooks without parameters (#11)
 * Handle non-code cells (#12)
+* Ignore magic commands in input notebooks (#14)
 
 ## Changes in 0.1.0